+++ prepline_general/api/general.py
@@ -19,7 +19,7 @@ from typing import Optional, Mapping, Iterator, Tuple
 import secrets
 from concurrent.futures import ThreadPoolExecutor
 from functools import partial
-from PyPDF2 import PdfReader, PdfWriter
+from pypdf import PdfReader, PdfWriter
 from unstructured.partition.api import partition_via_api
 from unstructured.partition.auto import partition
 from unstructured.staging.base import convert_to_isd, convert_to_dataframe