+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.0.27-dev0
+## 0.0.27
 
+* Update documentation and tests for filetypes to sync with partition.auto
+* Add support for .rst, .tsv, .xml
 * Move PYPDF2 to pypdf since PYPDF2 is deprecated
 
 ## 0.0.26